enforced `int` type for `get_bins`

+++ src/chainconsumer/helpers.py
@@ -43,7 +43,7 @@ def get_bins(chain: Chain) -> int:
     if chain.bins is not None:
         return chain.bins
     max_v = 35 if chain.smooth > 0 else 100
-    return max((max_v, np.floor(1.0 * np.power(chain.samples.shape[0] / chain.samples.shape[1], 0.25))))
+    return max((max_v, int(np.floor(1.0 * np.power(chain.samples.shape[0] / chain.samples.shape[1], 0.25)))))
 
 
 def get_smoothed_bins(
